moving config to app
